refactor analyzer screen sections
