+++ lib/screens/poker_analyzer_screen.dart
@@ -123,7 +123,9 @@ import 'package:poker_analyzer/plugins/plugin_loader.dart';
 import 'package:poker_analyzer/plugins/plugin.dart';
 import '../services/demo_playback_controller.dart';
 
-
+part 'poker_analyzer/board_widgets.dart';
+part 'poker_analyzer/playback_panel.dart';
+part 'poker_analyzer/action_controls.dart';
 
 class PokerAnalyzerScreen extends StatefulWidget {
   final SavedHand? initialHand;
@@ -6750,1596 +6752,12 @@ class _OpponentCardRowSection extends StatelessWidget {
   }
 }
 
-class _TableBackgroundSection extends StatelessWidget {
-  final double scale;
-
-  const _TableBackgroundSection({
-    required this.scale,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final screenSize = MediaQuery.of(context).size;
-    final tableWidth = screenSize.width * 0.9 * scale;
-
-    final tableHeight = tableWidth * 0.55;
-
-    return Center(
-      child: CustomPaint(
-        size: Size(tableWidth, tableHeight),
-        painter: PokerTablePainter(),
-      ),
-    );
-  }
 }
 
-
-class _BetStacksOverlaySection extends StatelessWidget {
-  final double scale;
-  final _PokerAnalyzerScreenState state;
-
-  const _BetStacksOverlaySection({
-    required this.scale,
-    required this.state,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final screenSize = MediaQuery.of(state.context).size;
-    final tableWidth = screenSize.width * 0.9;
-    final tableHeight = tableWidth * 0.55;
-    final centerX = screenSize.width / 2 + 10;
-    final centerY = screenSize.height / 2 -
-        TableGeometryHelper.centerYOffset(state.numberOfPlayers, scale);
-    final radiusMod = TableGeometryHelper.radiusModifier(state.numberOfPlayers);
-    final radiusX = (tableWidth / 2 - 60) * scale * radiusMod;
-    final radiusY = (tableHeight / 2 + 90) * scale * radiusMod;
-
-    final List<Widget> chips = [];
-    for (int i = 0; i < state.numberOfPlayers; i++) {
-      final index = (i + state._viewIndex()) % state.numberOfPlayers;
-      final invested =
-          state._stackService.getInvestmentForStreet(index, state.currentStreet);
-      if (invested > 0) {
-        final angle = 2 * pi * i / state.numberOfPlayers + pi / 2;
-        final dx = radiusX * cos(angle);
-        final dy = radiusY * sin(angle);
-        final bias = TableGeometryHelper.verticalBiasFromAngle(angle) * scale;
-
-        final start = Offset(centerX + dx, centerY + dy + bias + 92 * scale);
-        final pos = Offset.lerp(start, Offset(centerX, centerY), 0.15)!;
-
-        final chipScale = scale * 0.8;
-        chips.add(Positioned(
-          left: pos.dx - 8 * chipScale,
-          top: pos.dy - 8 * chipScale,
-          child: BetStackChips(amount: invested, scale: chipScale),
-        ));
-      }
-    }
-    return Stack(children: chips);
-  }
-}
-
-class _ActionBetStackOverlaySection extends StatelessWidget {
-  final double scale;
-  final _PokerAnalyzerScreenState state;
-
-  const _ActionBetStackOverlaySection({
-    required this.scale,
-    required this.state,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final screenSize = MediaQuery.of(state.context).size;
-    final tableWidth = screenSize.width * 0.9;
-    final tableHeight = tableWidth * 0.55;
-    final centerX = screenSize.width / 2 + 10;
-    final centerY = screenSize.height / 2 -
-        TableGeometryHelper.centerYOffset(state.numberOfPlayers, scale);
-    final radiusMod = TableGeometryHelper.radiusModifier(state.numberOfPlayers);
-    final radiusX = (tableWidth / 2 - 60) * scale * radiusMod;
-    final radiusY = (tableHeight / 2 + 90) * scale * radiusMod;
-
-    final chips = <Widget>[];
-    for (int i = 0; i < state.numberOfPlayers; i++) {
-      final index = (i + state._viewIndex()) % state.numberOfPlayers;
-      final amount = state._actionBetStacks[index];
-      final angle = 2 * pi * i / state.numberOfPlayers + pi / 2;
-      final dx = radiusX * cos(angle);
-      final dy = radiusY * sin(angle);
-      final bias = TableGeometryHelper.verticalBiasFromAngle(angle) * scale;
-      final start = Offset(centerX + dx, centerY + dy + bias + 92 * scale);
-      final pos = Offset.lerp(start, Offset(centerX, centerY), 0.15)!;
-      final chipScale = scale * 0.8;
-      chips.add(Positioned(
-        left: pos.dx - 8 * chipScale,
-        top: pos.dy - 8 * chipScale,
-        child: AnimatedSwitcher(
-          duration: const Duration(milliseconds: 300),
-          transitionBuilder: (child, animation) => FadeTransition(
-            opacity: animation,
-            child: ScaleTransition(scale: animation, child: child),
-          ),
-          child: amount != null
-              ? BetStackChips(
-                  key: ValueKey(amount),
-                  amount: amount,
-                  scale: chipScale,
-                )
-              : SizedBox(key: const ValueKey('empty'), height: 16 * chipScale),
-        ),
-      ));
-    }
-
-    return Stack(children: chips);
-  }
-}
-
-class _InvestedChipsOverlaySection extends StatelessWidget {
-  final double scale;
-  final _PokerAnalyzerScreenState state;
-
-  const _InvestedChipsOverlaySection({
-    required this.scale,
-    required this.state,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final screenSize = MediaQuery.of(state.context).size;
-    final tableWidth = screenSize.width * 0.9;
-    final tableHeight = tableWidth * 0.55;
-    final centerX = screenSize.width / 2 + 10;
-    final centerY = screenSize.height / 2 -
-        TableGeometryHelper.centerYOffset(state.numberOfPlayers, scale);
-    final radiusMod = TableGeometryHelper.radiusModifier(state.numberOfPlayers);
-    final radiusX = (tableWidth / 2 - 60) * scale * radiusMod;
-    final radiusY = (tableHeight / 2 + 90) * scale * radiusMod;
-
-    final List<Widget> chips = [];
-    for (int i = 0; i < state.numberOfPlayers; i++) {
-      final index = (i + state._viewIndex()) % state.numberOfPlayers;
-      final invested =
-          state._stackService.getInvestmentForStreet(index, state.currentStreet);
-      if (invested > 0) {
-        final angle = 2 * pi * i / state.numberOfPlayers + pi / 2;
-        final dx = radiusX * cos(angle);
-        final dy = radiusY * sin(angle);
-        final bias = TableGeometryHelper.verticalBiasFromAngle(angle) * scale;
-
-        final playerActions = state.actions
-            .where((a) =>
-                a.playerIndex == index && a.street == state.currentStreet)
-            .toList();
-        final lastAction =
-            playerActions.isNotEmpty ? playerActions.last : null;
-        final color =
-            ActionFormattingHelper.actionColor(lastAction?.action ?? 'bet');
-        final start =
-            Offset(centerX + dx, centerY + dy + bias + 92 * scale);
-        final end = Offset.lerp(start, Offset(centerX, centerY), 0.2)!;
-        final animate =
-            state._playbackManager.shouldAnimatePlayer(state.currentStreet, index);
-        chips.add(Positioned.fill(
-          child: BetChipsOnTable(
-            start: start,
-            end: end,
-            chipCount: (invested / 20).clamp(1, 5).round(),
-            color: color,
-            scale: scale,
-            animate: animate,
-          ),
-        ));
-      }
-    }
-    return Stack(children: chips);
-  }
-}
-
-class _ActionHistorySection extends StatelessWidget {
-  final ActionHistoryService actionHistory;
-  final Map<int, String> playerPositions;
-  final Set<int> expandedStreets;
-  final ValueChanged<int> onToggleStreet;
-  final void Function(int index, ActionEntry entry) onEdit;
-  final void Function(int index) onDelete;
-  final void Function(int, int) onReorder;
-  final bool isLocked;
-
-  const _ActionHistorySection({
-    required this.actionHistory,
-    required this.playerPositions,
-    required this.expandedStreets,
-    required this.onToggleStreet,
-    required this.onEdit,
-    required this.onDelete,
-    required this.onReorder,
-    required this.isLocked,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ActionHistoryOverlay(
-      actionHistory: actionHistory,
-      playerPositions: playerPositions,
-      expandedStreets: expandedStreets,
-      onToggleStreet: onToggleStreet,
-      onEdit: onEdit,
-      onDelete: onDelete,
-      onReorder: onReorder,
-      isLocked: isLocked,
-    );
-  }
-}
-
-
-class _HudOverlaySection extends StatelessWidget {
-  final String streetName;
-  final String potText;
-  final String stackText;
-  final String? sprText;
-
-  const _HudOverlaySection({
-    required this.streetName,
-    required this.potText,
-    required this.stackText,
-    this.sprText,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.topRight,
-      child: HudOverlay(
-        streetName: streetName,
-        potText: potText,
-        stackText: stackText,
-        sprText: sprText,
-      ),
-    );
-  }
-}
-
-class _BoardTransitionBusyIndicator extends StatelessWidget {
-  const _BoardTransitionBusyIndicator();
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned.fill(
-      child: ColoredBox(
-        color: Colors.black38,
-        child: const Center(
-          child: SizedBox(
-            width: 40,
-            height: 40,
-            child: CircularProgressIndicator(),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _PerspectiveSwitchButton extends StatelessWidget {
-  final bool isPerspectiveSwitched;
-  final VoidCallback onToggle;
-
-  const _PerspectiveSwitchButton({
-    required this.isPerspectiveSwitched,
-    required this.onToggle,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      top: 8,
-      right: 8,
-      child: TextButton(
-        onPressed: onToggle,
-        child: const Text(
-          '👁 Смотреть от лица',
-          style: TextStyle(color: Colors.white),
-        ),
-      ),
-    );
-  }
-}
-
-class _PlaybackControlsSection extends StatelessWidget {
-  final bool isPlaying;
-  final int playbackIndex;
-  final int actionCount;
-  final Duration elapsedTime;
-  final VoidCallback onPlay;
-  final VoidCallback onPause;
-  final VoidCallback onPlayAll;
-  final VoidCallback onStepBackward;
-  final VoidCallback onStepForward;
-  final VoidCallback onPlaybackReset;
-  final ValueChanged<double> onSeek;
-  final VoidCallback onReset;
-  final VoidCallback onBack;
-  final bool focusOnHero;
-  final ValueChanged<bool> onFocusChanged;
-  final bool backDisabled;
-  final bool disabled;
-
-  const _PlaybackControlsSection({
-    required this.isPlaying,
-    required this.playbackIndex,
-    required this.actionCount,
-    required this.elapsedTime,
-    required this.onPlay,
-    required this.onPause,
-    required this.onPlayAll,
-    required this.onStepBackward,
-    required this.onStepForward,
-    required this.onPlaybackReset,
-    required this.onSeek,
-    required this.onReset,
-    required this.onBack,
-    required this.focusOnHero,
-    required this.onFocusChanged,
-    this.backDisabled = false,
-    this.disabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final Color iconColor = disabled ? Colors.grey : Colors.white;
-
-    return Column(
-      children: [
-        Row(
-          children: [
-            IconButton(
-              icon: Icon(Icons.skip_previous, color: iconColor),
-              onPressed: disabled ? null : onStepBackward,
-            ),
-            IconButton(
-              icon: Icon(Icons.replay, color: iconColor),
-              onPressed: disabled ? null : onPlaybackReset,
-            ),
-            IconButton(
-              icon: Icon(
-                isPlaying ? Icons.pause : Icons.play_arrow,
-                color: iconColor,
-              ),
-              onPressed:
-                  disabled ? null : (isPlaying ? onPause : onPlay),
-            ),
-            IconButton(
-              icon: Icon(Icons.skip_next, color: iconColor),
-              onPressed: disabled ? null : onStepForward,
-            ),
-            Expanded(
-              child: Slider(
-                value: playbackIndex.toDouble(),
-                min: 0,
-                max: actionCount > 0 ? actionCount.toDouble() : 1,
-                onChanged: disabled ? null : onSeek,
-                inactiveColor: Colors.grey,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16.0),
-          child: Column(
-            children: [
-              LinearProgressIndicator(
-                value: actionCount > 0 ? playbackIndex / actionCount : 0.0,
-                backgroundColor: Colors.white24,
-                valueColor: AlwaysStoppedAnimation<Color>(
-                    Theme.of(context).colorScheme.secondary),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                "Step \$playbackIndex / \$actionCount" +
-                    (isPlaying ? " - " + _formatDuration(elapsedTime) : ""),
-                style: const TextStyle(color: Colors.white),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(height: 8),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            ElevatedButton(
-              onPressed: disabled ? null : (isPlaying ? onPause : onPlayAll),
-              child: Text(isPlaying ? "Pause" : "Play All"),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            ElevatedButton(
-              onPressed: disabled ? null : onStepBackward,
-              child: const Text('Step Back'),
-            ),
-            const SizedBox(width: 8),
-            ElevatedButton(
-              onPressed: disabled ? null : onStepForward,
-              child: const Text('Step Forward'),
-            ),
-          ],
-        ),
-        const SizedBox(height: 10),
-        TextButton(
-          onPressed: backDisabled ? null : onBack,
-          child: const Text('Назад'),
-        ),
-        const SizedBox(height: 10),
-        TextButton(
-          onPressed: onReset,
-          child: const Text('Сбросить раздачу'),
-        ),
-        SwitchListTile(
-          title: const Text('Focus on Hero', style: TextStyle(color: Colors.white)),
-          value: focusOnHero,
-          onChanged: disabled ? null : onFocusChanged,
-          activeColor: Colors.deepPurple,
-        ),
-      ],
-    );
-  }
-}
-
-class _SaveLoadControlsSection extends StatelessWidget {
-  final VoidCallback onSave;
-  final VoidCallback onLoadLast;
-  final VoidCallback onLoadByName;
-  final VoidCallback onExportLast;
-  final VoidCallback onExportAll;
-  final VoidCallback onImport;
-  final VoidCallback onImportAll;
-  final bool disabled;
-
-  const _SaveLoadControlsSection({
-    required this.onSave,
-    required this.onLoadLast,
-    required this.onLoadByName,
-    required this.onExportLast,
-    required this.onExportAll,
-    required this.onImport,
-    required this.onImportAll,
-    this.disabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final Color iconColor = disabled ? Colors.grey : Colors.white;
-    return Row(
-      children: [
-        IconButton(
-          icon: Icon(Icons.save, color: iconColor),
-          onPressed: disabled ? null : onSave,
-        ),
-        IconButton(
-          icon: Icon(Icons.folder_open, color: iconColor),
-          onPressed: disabled ? null : onLoadLast,
-        ),
-        IconButton(
-          icon: Icon(Icons.list, color: iconColor),
-          onPressed: disabled ? null : onLoadByName,
-        ),
-        IconButton(
-          icon: Icon(Icons.upload, color: iconColor),
-          onPressed: disabled ? null : onExportLast,
-        ),
-        IconButton(
-          icon: Icon(Icons.file_upload, color: iconColor),
-          onPressed: disabled ? null : onExportAll,
-        ),
-        IconButton(
-          icon: Icon(Icons.download, color: iconColor),
-          onPressed: disabled ? null : onImport,
-        ),
-        IconButton(
-          icon: Icon(Icons.file_download, color: iconColor),
-          onPressed: disabled ? null : onImportAll,
-        ),
-      ],
-    );
-  }
-}
-
-class _PlaybackAndHandControls extends StatelessWidget {
-  final bool isPlaying;
-  final int playbackIndex;
-  final int actionCount;
-  final Duration elapsedTime;
-  final VoidCallback onPlay;
-  final VoidCallback onPause;
-  final VoidCallback onPlayAll;
-  final VoidCallback onStepBackward;
-  final VoidCallback onStepForward;
-  final VoidCallback onPlaybackReset;
-  final ValueChanged<double> onSeek;
-  final VoidCallback onSave;
-  final VoidCallback onLoadLast;
-  final VoidCallback onLoadByName;
-  final VoidCallback onExportLast;
-  final VoidCallback onExportAll;
-  final VoidCallback onImport;
-  final VoidCallback onImportAll;
-  final VoidCallback onReset;
-  final VoidCallback onBack;
-  final bool focusOnHero;
-  final ValueChanged<bool> onFocusChanged;
-  final bool backDisabled;
-  final bool disabled;
-
-  const _PlaybackAndHandControls({
-    required this.isPlaying,
-    required this.playbackIndex,
-    required this.actionCount,
-    required this.elapsedTime,
-    required this.onPlay,
-    required this.onPause,
-    required this.onPlayAll,
-    required this.onStepBackward,
-    required this.onStepForward,
-    required this.onPlaybackReset,
-    required this.onSeek,
-    required this.onSave,
-    required this.onLoadLast,
-    required this.onLoadByName,
-    required this.onExportLast,
-    required this.onExportAll,
-    required this.onImport,
-    required this.onImportAll,
-    required this.onReset,
-    required this.onBack,
-    required this.focusOnHero,
-    required this.onFocusChanged,
-    this.backDisabled = false,
-    this.disabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        _SaveLoadControlsSection(
-          onSave: onSave,
-          onLoadLast: onLoadLast,
-          onLoadByName: onLoadByName,
-          onExportLast: onExportLast,
-          onExportAll: onExportAll,
-          onImport: onImport,
-          onImportAll: onImportAll,
-          disabled: disabled,
-        ),
-        const SizedBox(height: 10),
-        _PlaybackControlsSection(
-          isPlaying: isPlaying,
-          playbackIndex: playbackIndex,
-          actionCount: actionCount,
-          elapsedTime: elapsedTime,
-          onPlay: onPlay,
-          onPause: onPause,
-          onPlayAll: onPlayAll,
-          onStepBackward: onStepBackward,
-          onStepForward: onStepForward,
-          onPlaybackReset: onPlaybackReset,
-          onSeek: onSeek,
-          onReset: onReset,
-          onBack: onBack,
-          focusOnHero: focusOnHero,
-          onFocusChanged: onFocusChanged,
-          backDisabled: backDisabled,
-          disabled: disabled,
-        ),
-      ],
-    );
-  }
-}
 
 /// Collapsed view of action history with tabs for each street.
-class _CollapsibleActionHistorySection extends StatelessWidget {
-  final ActionHistoryService actionHistory;
-  final Map<int, String> playerPositions;
-  final int heroIndex;
 
-  const _CollapsibleActionHistorySection({
-    required this.actionHistory,
-    required this.playerPositions,
-    required this.heroIndex,
-  });
 
-  @override
-  Widget build(BuildContext context) {
-    return CollapsibleActionHistory(
-      actionHistory: actionHistory,
-      playerPositions: playerPositions,
-      heroIndex: heroIndex,
-    );
-  }
-}
-
-class _HandNotesSection extends StatelessWidget {
-  final TextEditingController commentController;
-  final TextEditingController tagsController;
-  final TextEditingController tournamentIdController;
-  final TextEditingController buyInController;
-  final TextEditingController prizePoolController;
-  final TextEditingController entrantsController;
-  final TextEditingController gameTypeController;
-
-  const _HandNotesSection({
-    required this.commentController,
-    required this.tagsController,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16.0),
-          child: TextField(
-            controller: commentController,
-            maxLines: 3,
-            style: const TextStyle(color: Colors.white),
-            decoration: const InputDecoration(
-              labelText: 'Комментарий к раздаче',
-              labelStyle: TextStyle(color: Colors.white),
-              filled: true,
-              fillColor: Colors.white12,
-              border: OutlineInputBorder(),
-            ),
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16.0),
-          child: TextField(
-            controller: tagsController,
-            style: const TextStyle(color: Colors.white),
-            decoration: const InputDecoration(
-              labelText: 'Теги',
-              labelStyle: TextStyle(color: Colors.white),
-              filled: true,
-              fillColor: Colors.white12,
-              border: OutlineInputBorder(),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _TournamentInfoSection extends StatefulWidget {
-  final TextEditingController idController;
-  final TextEditingController buyInController;
-  final TextEditingController prizePoolController;
-  final TextEditingController entrantsController;
-  final TextEditingController gameTypeController;
-
-  const _TournamentInfoSection({
-    required this.idController,
-    required this.buyInController,
-    required this.prizePoolController,
-    required this.entrantsController,
-    required this.gameTypeController,
-  });
-
-  @override
-  State<_TournamentInfoSection> createState() => _TournamentInfoSectionState();
-}
-
-class _TournamentInfoSectionState extends State<_TournamentInfoSection> {
-  bool _open = false;
-
-  bool get _allEmpty =>
-      widget.idController.text.isEmpty &&
-      widget.buyInController.text.isEmpty &&
-      widget.prizePoolController.text.isEmpty &&
-      widget.entrantsController.text.isEmpty &&
-      widget.gameTypeController.text.isEmpty;
-
-  Widget _summaryRow(String label, String value) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 2),
-      child: Text('$label: $value',
-          style: const TextStyle(color: Colors.white70)),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final summary = <Widget>[];
-    if (widget.idController.text.isNotEmpty) {
-      summary.add(_summaryRow('ID', widget.idController.text));
-    }
-    if (widget.buyInController.text.isNotEmpty) {
-      summary.add(_summaryRow('Buy-In', widget.buyInController.text));
-    }
-    if (widget.prizePoolController.text.isNotEmpty) {
-      summary.add(_summaryRow('Prize Pool', widget.prizePoolController.text));
-    }
-    if (widget.entrantsController.text.isNotEmpty) {
-      summary.add(_summaryRow('Entrants', widget.entrantsController.text));
-    }
-    if (widget.gameTypeController.text.isNotEmpty) {
-      summary.add(_summaryRow('Game', widget.gameTypeController.text));
-    }
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 4),
-      decoration: BoxDecoration(
-        color: Colors.black.withOpacity(0.4),
-        borderRadius: BorderRadius.circular(12),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          InkWell(
-            onTap: () => setState(() => _open = !_open),
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Row(
-                children: [
-                  const Expanded(
-                    child: Text('Tournament Info',
-                        style: TextStyle(
-                            color: Colors.white, fontWeight: FontWeight.bold)),
-                  ),
-                  AnimatedRotation(
-                    turns: _open ? 0.5 : 0,
-                    duration: const Duration(milliseconds: 200),
-                    child:
-                        const Icon(Icons.expand_more, color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          if (!_open && summary.isNotEmpty)
-            Padding(
-              padding:
-                  const EdgeInsets.symmetric(horizontal: 8.0, vertical: 4.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: summary,
-              ),
-            ),
-          ClipRect(
-            child: AnimatedAlign(
-              alignment: Alignment.topCenter,
-              duration: const Duration(milliseconds: 300),
-              heightFactor: _open ? 1 : 0,
-              child: Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 8.0, vertical: 4.0),
-                child: Column(
-                  children: [
-                    TextField(
-                      controller: widget.idController,
-                      style: const TextStyle(color: Colors.white),
-                      decoration: const InputDecoration(
-                        labelText: 'Tournament ID',
-                        labelStyle: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    TextField(
-                      controller: widget.buyInController,
-                      keyboardType: TextInputType.number,
-                      inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                      style: const TextStyle(color: Colors.white),
-                      decoration: const InputDecoration(
-                        labelText: 'Buy-In',
-                        labelStyle: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    TextField(
-                      controller: widget.prizePoolController,
-                      keyboardType: TextInputType.number,
-                      inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                      style: const TextStyle(color: Colors.white),
-                      decoration: const InputDecoration(
-                        labelText: 'Prize Pool',
-                        labelStyle: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    TextField(
-                      controller: widget.entrantsController,
-                      keyboardType: TextInputType.number,
-                      inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                      style: const TextStyle(color: Colors.white),
-                      decoration: const InputDecoration(
-                        labelText: 'Entrants',
-                        labelStyle: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    DropdownButtonFormField<String>(
-                      value: widget.gameTypeController.text.isEmpty
-                          ? null
-                          : widget.gameTypeController.text,
-                      decoration: const InputDecoration(
-                        labelText: 'Game Type',
-                        labelStyle: TextStyle(color: Colors.white),
-                      ),
-                      dropdownColor: Colors.grey[900],
-                      items: const [
-                        DropdownMenuItem(
-                            value: "Hold'em NL", child: Text("Hold'em NL")),
-                        DropdownMenuItem(value: 'Omaha PL', child: Text('Omaha PL')),
-                        DropdownMenuItem(value: 'Other', child: Text('Other')),
-                      ],
-                      onChanged: (v) =>
-                          setState(() => widget.gameTypeController.text = v ?? ''),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _StreetActionsSection extends StatelessWidget {
-  final int street;
-  final ActionHistoryService actionHistory;
-  final List<int> pots;
-  final Map<int, int> stackSizes;
-  final Map<int, String> playerPositions;
-  final PotSyncService potSync;
-  final List<ActionEntry> allActions;
-  final void Function(int, ActionEntry) onEdit;
-  final void Function(int) onDelete;
-  final void Function(int index, ActionEntry entry) onInsert;
-  final void Function(int) onDuplicate;
-  final void Function(int, int) onReorder;
-  final int? visibleCount;
-  final String Function(ActionEntry)? evaluateActionQuality;
-
-  const _StreetActionsSection({
-    required this.street,
-    required this.actionHistory,
-    required this.pots,
-    required this.stackSizes,
-    required this.playerPositions,
-    required this.potSync,
-    required this.allActions,
-    required this.onEdit,
-    required this.onDelete,
-    required this.onInsert,
-    required this.onDuplicate,
-    required this.onReorder,
-    this.visibleCount,
-    this.evaluateActionQuality,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final pot = pots[street];
-    final effStack = potSync.calculateEffectiveStackForStreet(
-        street, allActions, playerPositions.length);
-    final double? sprValue = pot > 0 ? effStack / pot : null;
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0),
-      child: StreetActionsList(
-        street: street,
-        actions: actionHistory.actionsForStreet(street,
-            collapsed: false),
-        pots: pots,
-        stackSizes: stackSizes,
-        playerPositions: playerPositions,
-        numberOfPlayers: playerPositions.length,
-        onEdit: onEdit,
-        onDelete: onDelete,
-        onInsert: onInsert,
-        onDuplicate: onDuplicate,
-        onReorder: onReorder,
-        visibleCount: visibleCount,
-        evaluateActionQuality: evaluateActionQuality,
-        sprValue: sprValue,
-      ),
-    );
-  }
-}
-
-class _HandHeaderSection extends StatelessWidget {
-  final String handName;
-  final int playerCount;
-  final String streetName;
-  final VoidCallback onEdit;
-
-  const _HandHeaderSection({
-    required this.handName,
-    required this.playerCount,
-    required this.streetName,
-    required this.onEdit,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(AppConstants.padding16),
-      child: Card(
-        color: AppColors.cardBackground,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(AppConstants.radius8),
-        ),
-        child: Padding(
-          padding: const EdgeInsets.all(12),
-          child: Row(
-            children: [
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      handName,
-                      style: const TextStyle(
-                        color: Colors.white,
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    const SizedBox(height: 4),
-                    Text(
-                      '${playerCount} players • $streetName',
-                      style: const TextStyle(color: Colors.white70),
-                    ),
-                  ],
-                ),
-              ),
-              IconButton(
-                icon: const Icon(Icons.edit, color: Colors.white),
-                onPressed: onEdit,
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _PlayerCountSelector extends StatelessWidget {
-  final int numberOfPlayers;
-  final Map<int, String> playerPositions;
-  final Map<int, PlayerType> playerTypes;
-  final ValueChanged<int>? onChanged;
-  final bool disabled;
-
-  const _PlayerCountSelector({
-    required this.numberOfPlayers,
-    required this.playerPositions,
-    required this.playerTypes,
-    this.onChanged,
-    this.disabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return DropdownButton<int>(
-      value: numberOfPlayers,
-      dropdownColor: Colors.black,
-      style: const TextStyle(color: Colors.white),
-      iconEnabledColor: Colors.white,
-      items: [
-        for (int i = 2; i <= 10; i++)
-          DropdownMenuItem(value: i, child: Text('Игроков: $i')),
-      ],
-      onChanged: disabled
-          ? null
-          : (value) {
-              if (value != null && onChanged != null) {
-                onChanged!(value);
-              }
-            },
-    );
-  }
-}
-
-class _AnalyzeButtonSection extends StatelessWidget {
-  final VoidCallback onAnalyze;
-
-  const _AnalyzeButtonSection({
-    required this.onAnalyze,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      onPressed: onAnalyze,
-      child: const Text('🔍 Проанализировать'),
-    );
-  }
-}
-
-class _PlayerEditorSection extends StatefulWidget {
-  final int initialStack;
-  final PlayerType initialType;
-  final bool isHeroSelected;
-  final CardModel? card1;
-  final CardModel? card2;
-  final bool disableCards;
-  final void Function(int, PlayerType, bool, CardModel?, CardModel?) onSave;
-
-  const _PlayerEditorSection({
-    required this.initialStack,
-    required this.initialType,
-    required this.isHeroSelected,
-    this.card1,
-    this.card2,
-    this.disableCards = false,
-    required this.onSave,
-  });
-
-  @override
-  State<_PlayerEditorSection> createState() => _PlayerEditorSectionState();
-}
-
-class _PlayerEditorSectionState extends State<_PlayerEditorSection> {
-  late TextEditingController _stackController;
-  late PlayerType _type;
-  late bool _isHero;
-  CardModel? _card1;
-  CardModel? _card2;
-
-  static const ranks = ['A', 'K', 'Q', 'J', 'T', '9', '8', '7', '6', '5', '4', '3', '2'];
-  static const suits = ['♠', '♥', '♦', '♣'];
-
-  @override
-  void initState() {
-    super.initState();
-    _stackController = TextEditingController(text: widget.initialStack.toString());
-    _type = widget.initialType;
-    _isHero = widget.isHeroSelected;
-    _card1 = widget.card1;
-    _card2 = widget.card2;
-  }
-
-  @override
-  void dispose() {
-    _stackController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      backgroundColor: Colors.grey[900],
-      title: const Text('Edit Player', style: TextStyle(color: Colors.white)),
-      content: SingleChildScrollView(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            TextField(
-              controller: _stackController,
-              keyboardType: TextInputType.number,
-              inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-              style: const TextStyle(color: Colors.white),
-              decoration: InputDecoration(
-                labelText: 'Stack',
-                labelStyle: const TextStyle(color: Colors.white70),
-                filled: true,
-                fillColor: Colors.white10,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-            ),
-            const SizedBox(height: 12),
-            DropdownButtonFormField<String>(
-              value: _type.name,
-              dropdownColor: Colors.grey[900],
-              decoration: const InputDecoration(
-                labelText: 'Type',
-                labelStyle: TextStyle(color: Colors.white70),
-              ),
-              items: const [
-                DropdownMenuItem(value: 'standard', child: Text('standard')),
-                DropdownMenuItem(value: 'nit', child: Text('nit')),
-                DropdownMenuItem(value: 'fish', child: Text('fish')),
-                DropdownMenuItem(value: 'maniac', child: Text('maniac')),
-                DropdownMenuItem(value: 'shark', child: Text('shark')),
-                DropdownMenuItem(value: 'station', child: Text('station')),
-              ],
-              onChanged: (v) {
-                if (v != null) {
-                  lockService.safeSetState(this, () =>
-                      _type = PlayerType.values.firstWhere(
-                        (e) => e.name == v,
-                        orElse: () => PlayerType.unknown,
-                      ));
-                }
-              },
-            ),
-            const SizedBox(height: 12),
-            SwitchListTile(
-              value: _isHero,
-              title: const Text('Hero', style: TextStyle(color: Colors.white)),
-              onChanged: (v) => context.findAncestorStateOfType<_PokerAnalyzerScreenState>()?.lockService.safeSetState(this, () => _isHero = v),
-              activeColor: Colors.orange,
-            ),
-            const SizedBox(height: 12),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Column(
-                  children: [
-                    DropdownButton<String>(
-                      value: _card1?.rank,
-                      hint: const Text('Rank', style: TextStyle(color: Colors.white54)),
-                      dropdownColor: Colors.grey[900],
-                      items: ranks
-                          .map((r) => DropdownMenuItem(value: r, child: Text(r)))
-                          .toList(),
-                      onChanged: widget.disableCards
-                          ? null
-                          : (v) => lockService.safeSetState(this, () => _card1 = v != null && _card1 != null
-                              ? CardModel(rank: v, suit: _card1!.suit)
-                              : (v != null ? CardModel(rank: v, suit: suits.first) : null)),
-                    ),
-                    DropdownButton<String>(
-                      value: _card1?.suit,
-                      hint: const Text('Suit', style: TextStyle(color: Colors.white54)),
-                      dropdownColor: Colors.grey[900],
-                      items: suits
-                          .map((s) => DropdownMenuItem(value: s, child: Text(s)))
-                          .toList(),
-                      onChanged: widget.disableCards
-                          ? null
-                          : (v) => lockService.safeSetState(this, () => _card1 = v != null && _card1 != null
-                              ? CardModel(rank: _card1!.rank, suit: v)
-                              : (v != null ? CardModel(rank: ranks.first, suit: v) : null)),
-                    ),
-                  ],
-                ),
-                Column(
-                  children: [
-                    DropdownButton<String>(
-                      value: _card2?.rank,
-                      hint: const Text('Rank', style: TextStyle(color: Colors.white54)),
-                      dropdownColor: Colors.grey[900],
-                      items: ranks
-                          .map((r) => DropdownMenuItem(value: r, child: Text(r)))
-                          .toList(),
-                      onChanged: widget.disableCards
-                          ? null
-                          : (v) => lockService.safeSetState(this, () => _card2 = v != null && _card2 != null
-                              ? CardModel(rank: v, suit: _card2!.suit)
-                              : (v != null ? CardModel(rank: v, suit: suits.first) : null)),
-                    ),
-                    DropdownButton<String>(
-                      value: _card2?.suit,
-                      hint: const Text('Suit', style: TextStyle(color: Colors.white54)),
-                      dropdownColor: Colors.grey[900],
-                      items: suits
-                          .map((s) => DropdownMenuItem(value: s, child: Text(s)))
-                          .toList(),
-                      onChanged: widget.disableCards
-                          ? null
-                          : (v) => lockService.safeSetState(this, () => _card2 = v != null && _card2 != null
-                              ? CardModel(rank: _card2!.rank, suit: v)
-                              : (v != null ? CardModel(rank: ranks.first, suit: v) : null)),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        TextButton(
-          onPressed: () {
-            final stack = int.tryParse(_stackController.text) ?? widget.initialStack;
-            widget.onSave(stack, _type, _isHero, _card1, _card2);
-            Navigator.pop(context);
-          },
-          child: const Text('OK'),
-        ),
-      ],
-    );
-  }
-}
-
-class _HandEditorSection extends StatelessWidget {
-  final ActionHistoryService actionHistory;
-  final Map<int, String> playerPositions;
-  final int heroIndex;
-  final TextEditingController commentController;
-  final TextEditingController tagsController;
-  final int currentStreet;
-  final List<int> pots;
-  final Map<int, int> stackSizes;
-  final void Function(int, ActionEntry) onEdit;
-  final void Function(int) onDelete;
-  final int? visibleCount;
-  final String Function(ActionEntry)? evaluateActionQuality;
-  final VoidCallback onAnalyze;
-
-  const _HandEditorSection({
-    required this.actionHistory,
-    required this.playerPositions,
-    required this.heroIndex,
-    required this.commentController,
-    required this.tagsController,
-    required this.tournamentIdController,
-    required this.buyInController,
-    required this.prizePoolController,
-    required this.entrantsController,
-    required this.gameTypeController,
-    required this.currentStreet,
-    required this.pots,
-    required this.stackSizes,
-    required this.onEdit,
-    required this.onDelete,
-    required this.visibleCount,
-    required this.evaluateActionQuality,
-    required this.onAnalyze,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        _CollapsibleActionHistorySection(
-          actionHistory: _actionHistory,
-          playerPositions: playerPositions,
-          heroIndex: heroIndex,
-        ),
-        Expanded(
-          child: SingleChildScrollView(
-            child: Column(
-              children: [
-                _HandNotesSection(
-                    commentController: commentController,
-                    tagsController: tagsController),
-                _TournamentInfoSection(
-                  idController: tournamentIdController,
-                  buyInController: buyInController,
-                  prizePoolController: prizePoolController,
-                  entrantsController: entrantsController,
-                  gameTypeController: gameTypeController,
-                ),
-                _StreetActionsSection(
-                  street: currentStreet,
-                  actionHistory: _actionHistory,
-                  pots: pots,
-                  stackSizes: stackSizes,
-                  playerPositions: playerPositions,
-                  potSync: _potSync,
-                  allActions: actions,
-                  onEdit: onEdit,
-                  onDelete: onDelete,
-                  onInsert: _insertAction,
-                  onDuplicate: _duplicateAction,
-                  onReorder: _reorderAction,
-                  visibleCount: visibleCount,
-                  evaluateActionQuality: evaluateActionQuality,
-                ),
-                const SizedBox(height: 10),
-                _AnalyzeButtonSection(onAnalyze: onAnalyze),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _RevealAllCardsButton extends StatelessWidget {
-  final bool showAllRevealedCards;
-  final VoidCallback onToggle;
-
-  const _RevealAllCardsButton({
-    required this.showAllRevealedCards,
-    required this.onToggle,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: 8.0),
-        child: ElevatedButton(
-          onPressed: onToggle,
-          child: const Text('Показать все карты'),
-        ),
-      ),
-    );
-  }
-}
-
-class _FinishHandButtonOverlay extends StatelessWidget {
-  final VoidCallback onPressed;
-  final bool disabled;
-  final bool visible;
-
-  const _FinishHandButtonOverlay({
-    required this.onPressed,
-    required this.disabled,
-    required this.visible,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: 100.0),
-        child: AnimatedOpacity(
-          opacity: visible ? 1.0 : 0.0,
-          duration: const Duration(milliseconds: 300),
-          child: ElevatedButton(
-            onPressed: disabled ? null : onPressed,
-            child: const Text('Завершить раздачу'),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _NextHandButtonOverlay extends StatelessWidget {
-  final VoidCallback onPressed;
-
-  const _NextHandButtonOverlay({required this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: 40.0),
-        child: Container(
-          decoration: BoxDecoration(
-            boxShadow: [
-              BoxShadow(
-                color: Colors.yellowAccent.withOpacity(0.6),
-                blurRadius: 20,
-                spreadRadius: 5,
-              ),
-            ],
-          ),
-          child: ElevatedButton(
-            onPressed: onPressed,
-            style: ElevatedButton.styleFrom(
-              padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 16),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(30),
-              ),
-              textStyle: const TextStyle(fontSize: 20),
-            ),
-            child: const Text('Next Hand'),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _HandCompleteOverlay extends StatelessWidget {
-  final bool visible;
-
-  const _HandCompleteOverlay({required this.visible});
-
-  @override
-  Widget build(BuildContext context) {
-    return IgnorePointer(
-      child: AnimatedOpacity(
-        opacity: visible ? 1.0 : 0.0,
-        duration: const Duration(milliseconds: 500),
-        child: Center(
-          child: Container(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-            decoration: BoxDecoration(
-              color: Colors.black.withOpacity(0.8),
-              borderRadius: BorderRadius.circular(20),
-            ),
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: const [
-                Icon(Icons.check_circle, color: Colors.greenAccent, size: 28),
-                SizedBox(width: 8),
-                Text(
-                  'Раздача завершена',
-                  style: TextStyle(color: Colors.white, fontSize: 20),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ReplayDemoButtonOverlay extends StatelessWidget {
-  final VoidCallback onPressed;
-  final bool visible;
-
-  const _ReplayDemoButtonOverlay({
-    required this.onPressed,
-    required this.visible,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: 40.0),
-        child: AnimatedOpacity(
-          opacity: visible ? 1.0 : 0.0,
-          duration: const Duration(milliseconds: 500),
-          child: ElevatedButton(
-            onPressed: onPressed,
-            style: ElevatedButton.styleFrom(
-              padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 16),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(30),
-              ),
-              textStyle: const TextStyle(fontSize: 20),
-            ),
-            child: const Text('Replay Demo'),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _FoldAllButtonOverlay extends StatelessWidget {
-  final VoidCallback onPressed;
-
-  const _FoldAllButtonOverlay({required this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return Positioned(
-      bottom: 16,
-      right: 16,
-      child: FloatingActionButton.extended(
-        heroTag: 'foldAllFab',
-        backgroundColor: Colors.red,
-        onPressed: onPressed,
-        label: const Text('Fold All'),
-      ),
-    );
-  }
-}
-
-class _HandProgressIndicator extends StatelessWidget {
-  final int step;
-
-  const _HandProgressIndicator({required this.step});
-
-  @override
-  Widget build(BuildContext context) {
-    const labels = ['Игроки', 'Карты', 'Действия', 'Шоудаун'];
-    const icons = [
-      Icons.people,
-      Icons.style,
-      Icons.list_alt,
-      Icons.flag,
-    ];
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: List.generate(labels.length, (i) {
-          final active = step >= i;
-          return Expanded(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                AnimatedContainer(
-                  duration: const Duration(milliseconds: 300),
-                  padding: const EdgeInsets.all(6),
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    color: active ? Colors.blueAccent : Colors.white12,
-                  ),
-                  child: Icon(
-                    icons[i],
-                    size: 16,
-                    color: Colors.white,
-                  ),
-                ),
-                const SizedBox(height: 4),
-                Text(
-                  labels[i],
-                  textAlign: TextAlign.center,
-                  style: TextStyle(
-                    color: active ? Colors.white : Colors.white54,
-                    fontSize: 12,
-                  ),
-                ),
-              ],
-            ),
-          );
-        }),
-      ),
-    );
-  }
-}
-
-class _TotalPotTracker extends StatelessWidget {
-  final PotSyncService potSync;
-  final int currentStreet;
-  final List<int> sidePots;
-
-  const _TotalPotTracker({
-    required this.potSync,
-    required this.currentStreet,
-    required this.sidePots,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final currentPot =
-        currentStreet < potSync.pots.length ? potSync.pots[currentStreet] : 0;
-    final sideTotal = sidePots.fold<int>(0, (p, e) => p + e);
-    final totalPot = currentPot + sideTotal;
-    if (totalPot <= 0) {
-      return const SizedBox.shrink();
-    }
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 4.0),
-      child: Text(
-        'Total Pot: ${ActionFormattingHelper.formatAmount(totalPot)}',
-        style: const TextStyle(
-          color: AppColors.accent,
-          fontWeight: FontWeight.bold,
-          fontSize: 16,
-        ),
-      ),
-    );
-  }
-}
-
-class _InfoBanner extends StatelessWidget {
-  final String message;
-
-  const _InfoBanner({required this.message});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4.0),
-      child: Container(
-        width: double.infinity,
-        padding: const EdgeInsets.all(12.0),
-        decoration: BoxDecoration(
-          color: AppColors.cardBackground,
-          borderRadius: BorderRadius.circular(AppConstants.radius8),
-        ),
-        child: Text(
-          message,
-          style: const TextStyle(color: Colors.white70),
-        ),
-      ),
-    );
-  }
-}
 
 class StreetActionInputWidget extends StatefulWidget {
   final int currentStreet;
